% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvnRnd.R
\name{MvnRnd}
\alias{MvnRnd}
\title{Simulations with Exact Means and Covariances}
\usage{
MvnRnd(M, S, J)
}
\arguments{
\item{M}{A \code{1xN} matrix with the location parameter of each invariant (asset).}

\item{S}{A \code{NxN} matrix with the dispersion matrix of the invariants (assets).}

\item{J}{A numeric scalar with the desired number of scenarios.}
}
\value{
A J x N matrix with the simulated time-series.
}
\description{
This function generate scenarios from elliptical distributions in which the
sample moments match the populational moments.
}
\examples{
rets  <- diff(log(EuStockMarkets))
mu    <- colMeans(rets)
sigma <- stats::cov(rets)
MvnRnd(M = mu, S = sigma, J = 50)
}
\references{
Meucci, Attilio, Simulations with Exact Means and Covariances (June 7, 2009). Available at SSRN: \url{https://ssrn.com/abstract=1415699} or \url{http://dx.doi.org/10.2139/ssrn.1415699}.

Attilio Meucci (2020). Simulations with Exact Means and Covariances, \href{https://www.mathworks.com/matlabcentral/fileexchange/24416-simulations-with-exact-means-and-covariances}{MATLAB Central File Exchange}. Retrieved October 11, 2020.
}
